+++ agent/agentexec/cli_unix.go
@@ -1,3 +1,6 @@
+//go:build linux
+// +build linux
+
 package agentexec
 
 import (
@@ -12,11 +15,6 @@ import (
 
 	"golang.org/x/sys/unix"
 	"golang.org/x/xerrors"
-)
-
-const (
-	EnvProcOOMScore  = "CODER_PROC_OOM_SCORE"
-	EnvProcNiceScore = "CODER_PROC_NICE_SCORE"
 )
 
 // CLI runs the agent-exec command. It should only be called by the cli package.
@@ -126,31 +124,4 @@ func oomScoreAdj(pid int) (int, error) {
 
 func writeOOMScoreAdj(pid int, score int) error {
 	return os.WriteFile(fmt.Sprintf("/proc/%d/oom_score_adj", pid), []byte(fmt.Sprintf("%d", score)), 0o600)
-}
-
-// envValInt searches for a key in a list of environment variables and parses it to an int.
-// If the key is not found or cannot be parsed, returns 0 and false.
-func envValInt(env []string, key string) (int, bool) {
-	val, ok := envVal(env, key)
-	if !ok {
-		return 0, false
-	}
-
-	i, err := strconv.Atoi(val)
-	if err != nil {
-		return 0, false
-	}
-	return i, true
-}
-
-// envVal searches for a key in a list of environment variables and returns its value.
-// If the key is not found, returns empty string and false.
-func envVal(env []string, key string) (string, bool) {
-	prefix := key + "="
-	for _, e := range env {
-		if strings.HasPrefix(e, prefix) {
-			return strings.TrimPrefix(e, prefix), true
-		}
-	}
-	return "", false
 }
