+++ agent/agentexec/exec.go
@@ -6,6 +6,8 @@ import (
 	"os/exec"
 	"path/filepath"
 	"runtime"
+	"strconv"
+	"strings"
 
 	"golang.org/x/xerrors"
 )
@@ -13,7 +15,9 @@ import (
 const (
 	// EnvProcPrioMgmt is the environment variable that determines whether
 	// we attempt to manage process CPU and OOM Killer priority.
-	EnvProcPrioMgmt = "CODER_PROC_PRIO_MGMT"
+	EnvProcPrioMgmt  = "CODER_PROC_PRIO_MGMT"
+	EnvProcOOMScore  = "CODER_PROC_OOM_SCORE"
+	EnvProcNiceScore = "CODER_PROC_NICE_SCORE"
 )
 
 // CommandContext returns an exec.Cmd that calls "coder agent-exec" prior to exec'ing
@@ -39,4 +43,31 @@ func CommandContext(ctx context.Context, cmd string, args ...string) (*exec.Cmd,
 
 	args = append([]string{"agent-exec", cmd}, args...)
 	return exec.CommandContext(ctx, bin, args...), nil
+}
+
+// envValInt searches for a key in a list of environment variables and parses it to an int.
+// If the key is not found or cannot be parsed, returns 0 and false.
+func envValInt(env []string, key string) (int, bool) {
+	val, ok := envVal(env, key)
+	if !ok {
+		return 0, false
+	}
+
+	i, err := strconv.Atoi(val)
+	if err != nil {
+		return 0, false
+	}
+	return i, true
+}
+
+// envVal searches for a key in a list of environment variables and returns its value.
+// If the key is not found, returns empty string and false.
+func envVal(env []string, key string) (string, bool) {
+	prefix := key + "="
+	for _, e := range env {
+		if strings.HasPrefix(e, prefix) {
+			return strings.TrimPrefix(e, prefix), true
+		}
+	}
+	return "", false
 }
